[REF] mrp_bom_hierarchy with OCA guidelines
