+++ mrp_bom_hierarchy/README.rst
@@ -7,7 +7,7 @@ MRP BoM Hierarchy
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:f82342db68fac74511be10345e468d4de88b2c5cd1040c7bda76482629ac03c4
+   !! source digest: sha256:0056b6f5f7c205257b3021d59fc6e40b3ad4143fa7c916ad432a4dbcc0f101f9
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -68,6 +68,7 @@ Contributors
 
 * Jordi Ballester Alomar <jordi.ballester@forgeflow.com>
 * Miquel Raïch Regué <miquel.raich@forgeflow.com>
+* Quentin Dupont <quentin.dupont@grap.coop>
 
 Maintainers
 ~~~~~~~~~~~
