+++ mrp_bom_hierarchy/__manifest__.py
@@ -6,12 +6,13 @@
     "summary": "Make it easy to navigate through BoM hierarchy.",
     "version": "16.0.1.0.1",
     "author": "ForgeFlow, Odoo Community Association (OCA)",
+    "contributors": "Grap",
     "category": "Manufacturing",
     "depends": ["mrp"],
     "website": "https://github.com/OCA/manufacture",
     "license": "AGPL-3",
     "data": [
-        "view/mrp.xml",
+        "view/mrp_bom.xml",
     ],
     "installable": True,
     "auto_install": False,
