+++ mrp_bom_hierarchy/readme/CONTRIBUTORS.rst
@@ -1,2 +1,3 @@
 * Jordi Ballester Alomar <jordi.ballester@forgeflow.com>
 * Miquel Raïch Regué <miquel.raich@forgeflow.com>
+* Quentin Dupont <quentin.dupont@grap.coop>
